ENG-6426 use sitekey to get the Advanced key
