+++ NeuroID/src/main/java/com/neuroid/tracker/models/AdvancedDeviceKey.kt
@@ -1,3 +1,3 @@
-package com.neuroid.tracker.service
+package com.neuroid.tracker.models
 
 data class AdvancedDeviceKey(var status: String, var key: String)
