+++ NeuroID/src/main/java/com/neuroid/tracker/utils/GsonAdvMapper.kt
@@ -1,6 +1,7 @@
-package com.neuroid.tracker.service
+package com.neuroid.tracker.utils
 
 import com.google.gson.Gson
+import com.neuroid.tracker.models.AdvancedDeviceKey
 
 class GsonAdvMapper {
     fun getKey(data: String): AdvancedDeviceKey =
